IColumn::getPermutation() and IColumn::permutate() functions, "limit" parameter: change type UInt64 -> size_t.

+++ dbms/src/Columns/ColumnAggregateFunction.h
@@ -163,7 +163,7 @@ public:
 
     ColumnPtr filter(const Filter & filter, ssize_t result_size_hint) const override;
 
-    ColumnPtr permute(const Permutation & perm, UInt64 limit) const override;
+    ColumnPtr permute(const Permutation & perm, size_t limit) const override;
 
     ColumnPtr index(const IColumn & indexes, size_t limit) const override;
 
@@ -181,7 +181,7 @@ public:
         return 0;
     }
 
-    void getPermutation(bool reverse, UInt64 limit, int nan_direction_hint, Permutation & res) const override;
+    void getPermutation(bool reverse, size_t limit, int nan_direction_hint, Permutation & res) const override;
 
     /** More efficient manipulation methods */
     Container & getData()
