+++ dbms/src/Columns/ColumnDecimal.h
@@ -101,7 +101,7 @@ public:
     const char * deserializeAndInsertFromArena(const char * pos) override;
     void updateHashWithValue(size_t n, SipHash & hash) const override;
     int compareAt(size_t n, size_t m, const IColumn & rhs_, int nan_direction_hint) const override;
-    void getPermutation(bool reverse, UInt64 limit, int nan_direction_hint, IColumn::Permutation & res) const override;
+    void getPermutation(bool reverse, size_t limit, int nan_direction_hint, IColumn::Permutation & res) const override;
 
     MutableColumnPtr cloneResized(size_t size) const override;
 
@@ -116,7 +116,7 @@ public:
     bool isDefaultAt(size_t n) const override { return data[n] == 0; }
 
     ColumnPtr filter(const IColumn::Filter & filt, ssize_t result_size_hint) const override;
-    ColumnPtr permute(const IColumn::Permutation & perm, UInt64 limit) const override;
+    ColumnPtr permute(const IColumn::Permutation & perm, size_t limit) const override;
     ColumnPtr index(const IColumn & indexes, size_t limit) const override;
 
     template <typename Type>
@@ -144,7 +144,7 @@ protected:
     UInt32 scale;
 
     template <typename U>
-    void permutation(bool reverse, UInt64 limit, PaddedPODArray<U> & res) const
+    void permutation(bool reverse, size_t limit, PaddedPODArray<U> & res) const
     {
         size_t s = data.size();
         res.resize(s);
